+++ script/interactions/AddressLoader.sol
@@ -52,10 +52,6 @@ library AddressLoader {
         return 0x0165878A594ca255338adfa4d48449f69242Eb8F;
     }
 
-    function getAccumulator() internal pure returns (address) {
-        return 0xa513E6E4b8f2a923D98304ec87F64353C4D5C853;
-    }
-
     function getPhlimbo() internal pure returns (address) {
         return 0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6;
     }
@@ -89,7 +85,6 @@ library AddressLoader {
         console.log("YieldStrategyUSDT:", getYieldStrategyUSDT());
         console.log("YieldStrategyUSDS:", getYieldStrategyUSDS());
         console.log("PhusdStableMinter:", getMinter());
-        console.log("StableYieldAccumulator:", getAccumulator());
         console.log("PhlimboEA:", getPhlimbo());
         console.log("===================================");
     }